Fixed issue with Pagination where navigating to the first page caused a 404
The pagination code used an if statement to determine whether to head to the site's baseurl when the user was look at page 2 and wanting to view the posts on the main page.
This left off the the trailing slash, which would cause a 404 when navigating from the second page to the first page.